Options for input, displaying requested info, etc.
In getCityList func: added necessary checks for hiding/showing options' ul tag; display state of the city, if there is; calls getWeather func and passes coordinates of the selected option's city. Need to fix a bug: when typed fast in input, funciton displays all options for, for example, previous letter, and for next letter, so there could be many, many options.
Separated displaying received information logic from getWeather func to displayWeather func. In getWeather func: removed clearing previous information, now it doesn't change anything when clicked button with clear input; make get request by lat and lon(if known) or by city information(if no coordinates).
Need to fix a bug: if input's suggestion/option is clicked, getWeather func is called with coordinates(which are taken from the same response as a city info in option) as parameters. getWeather then make a request by lat and lon, but api returns a response with a different city name. It's probably cus it's a city, in which could be 2 locations - city and it's center(or some location inside that city). Previously the displayed location was taken from response. Now need to display it how it was selected, but without a state, if it is there.